SimpleChatTCRV:UICleanup+: Fix oversight wrt Auto ObjPropsEdit
Had forgotten to use ${} for variable within a literal template
string.

Had forgotten to use the full hierarchy of props when setting the
id of elements, which is useful to map a label to a button and so.
This would have led to the wrong button being clicked when a label
with duplicated id was clicked. Ie if two different objects within
a object, have properties/fields with the same name, then previously
this would have led to a bunch of labels and buttons or so having
the same id, bcas their hierarchy was not being accounted for. This
is fixed now.