SimpleChatTCRV:DivStreams: Why missing - still a mystry
Sometimes the session specific DivStream seems to be missing from
DivChat. Explicitly spliting the remove and append just for the
heck of it didnt help. So also have used try-catch-finally to
ensure it gets added back even if its not already in there, by
not allowing removeChild failure to bypass append etal.

Retaining this here just for the heck of it. WIll be switching
to the appendChild only flow (given it automatically moves the
node, as needed).