SimpleSallap:SimpleProxy:Make Config dataclass driven - p1
Instead of maintaining the config and some of the runtime states
identified as gMe as a generic literal dictionary which grows at
runtime with fields as required, try create it as a class of classes.

Inturn use dataclass annotation to let biolerplate code get auto
generated.

A config module created with above, however remaining part of the
code not yet updated to work with this new structure.

process_args and load_config moved into the new Config class.