SimpleSallap:SimpleMCP:TCWeb:XMLFiltered initial go wrt new flow
Also remember to picks the tagDropREs from passed args object and
not from got http header.

Even TCHtmlText updated to get the tags to drop from passed args
object and not got http header. And inturn allow ai to pass this
optional arg, as it sees fit in co-ordination with user.