SimpleSallap:SimpleMCP: Cleanup, Readme updates
Update the documentation a bit wrt switch from simpleproxy to
simplemcp with mcp-ish kind of handshake between the chat client
and simplemcp.

Rename proxyUrl and related to mcpServerUrl and mcpServerAuth.
Now include the path in the url itself, given that in future,
we may want to allow the chat client logic to handshake with
other mcp servers, which may expose their services through a
different path or so.

Drop SearchEngine related config entries from chat session
settings, given that its now controlled directly in SimpleMCP.