eval-callback : add support for saving logits
This commit adds support for saving logits to files in the evaluation
callback example. Two files are stored which is a binary file and a text
file for manual inspection.

Two options have been added to this example:
```console
----- example-specific params -----

--save-logits                           save final logits to files for verification (default: false)
--logits-output-dir PATH                directory for saving logits output files (default: data)
```

This motivation for this change (and follow up changes) is to replace
llama-logits in examples/model-conversion, which stores logits so they
can be compared to the original models logits.

Future commits will add more of the features that are currently in
llama-logits, like printing the prompt and token ids, and also enhance
this to also store the logits and token ids so that they can also be
compared as part of the verification process.